requirements.txt
PyPDF2 == 3.0.1
pytesseract == 0.3.13
pdfminer.six == 20240706
fitx == 1.0.1
fitz == 0.0.1.dev2
PyMuPDF == 1.24.10